Move `current_cycle` logic to `CalcStack`
Summary:
Deterministic cycle breaking will require this, so let's move it out
of `debugging.rs`

Reviewed By: ndmitchell

Differential Revision: D76948139

fbshipit-source-id: 888faefc096564811606ae543343d40066bf6fd9
